feat: Improve UI and functionality of progress and comments components
- Add 'height' prop to Next13ProgressBar component to set the height of
the progress bar to 4 pixels
- Update likeCount state in Comment component to include a 'disabled'
property and handle increase and decrease of like count
- Render avatar images using the Avatar component from the MUI Joy
library and handle rendering of contributor avatars when there are more
than 4 contributors
- Update Tech Stack section in Project component to add tooltips
displaying the names of the technologies
- Use Tooltip component from Ant Design library to add tooltips to the
stack icons in the Tech Stack section
- Use the Avatar component from the MUI Joy library to render user
avatars in Question component
- Improve styling and avatar rendering logic in both the Comment and
Question components
- Fix missing closing quote in the description of the 'generateMetadata'
function in questions.ts file

This commit improves the visual appearance and user experience of the
progress and comments components by adding a height prop to the progress
bar, updating the likeCount state with a disabled property, rendering
avatar images using the Avatar component, and fixing styling and avatar
rendering issues. It also enhances the Tech Stack section in the Project
component by adding tooltips to the stack icons and uses the Avatar
component to render user avatars in the Question component. These
changes improve the overall usability and aesthetics of the components.

Resolves: #42, #53